requirements-test.txt
nose
pyhamcrest
coverage
